+++ src/server/components/dashboard/relatedLinks/handlers/index.ts
@@ -1,2 +1,3 @@
 export * as edit from "./edit";
 export * as confirm from "./confirm";
+export { handleRelatedLinkIndexParam } from "./relatedLinkIndexParam";
